Moved roomkit api code over for server side execution with client interaction using subscribe/subscriptions architecture.
