+++ client/lib/subscribe.js
@@ -5,6 +5,7 @@ Template.index.events({
     'submit form': function (event) {
         event.preventDefault();
         console.log("Ping");
-        Meteor.call('pong', ("Pong"));
+        Meteor.call('login');
+        Meteor.call('get_volume');
     }
 });
